# internal/model/auth_model.go
package model

type LoginRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

type RegisterRequest struct {
	FirstName   string `json:"first_name" validate:"required"`
	LastName    string `json:"last_name" validate:"required"`
	Email       string `json:"email" validate:"required,email"`
	PhonePrefix string `json:"phone_prefix" validate:"required"`
	PhoneNumber string `json:"phone_number" validate:"required,numeric"`
	Password    string `json:"password" validate:"required,min=8"`
	RoleID      string `json:"role_id" validate:"required"`
}

type RefreshRequest struct {
	RefreshToken string `json:"refresh_token" validate:"required"`
}

type AuthResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}